+++ src/main/java/com/dogeyes/zyf/pojo/MovieExample.java
@@ -786,52 +786,52 @@ public class MovieExample {
             return (Criteria) this;
         }
 
-        public Criteria andRateEqualTo(BigDecimal value) {
+        public Criteria andRateEqualTo(Float value) {
             addCriterion("rate =", value, "rate");
             return (Criteria) this;
         }
 
-        public Criteria andRateNotEqualTo(BigDecimal value) {
+        public Criteria andRateNotEqualTo(Float value) {
             addCriterion("rate <>", value, "rate");
             return (Criteria) this;
         }
 
-        public Criteria andRateGreaterThan(BigDecimal value) {
+        public Criteria andRateGreaterThan(Float value) {
             addCriterion("rate >", value, "rate");
             return (Criteria) this;
         }
 
-        public Criteria andRateGreaterThanOrEqualTo(BigDecimal value) {
+        public Criteria andRateGreaterThanOrEqualTo(Float value) {
             addCriterion("rate >=", value, "rate");
             return (Criteria) this;
         }
 
-        public Criteria andRateLessThan(BigDecimal value) {
+        public Criteria andRateLessThan(Float value) {
             addCriterion("rate <", value, "rate");
             return (Criteria) this;
         }
 
-        public Criteria andRateLessThanOrEqualTo(BigDecimal value) {
+        public Criteria andRateLessThanOrEqualTo(Float value) {
             addCriterion("rate <=", value, "rate");
             return (Criteria) this;
         }
 
-        public Criteria andRateIn(List<BigDecimal> values) {
+        public Criteria andRateIn(List<Float> values) {
             addCriterion("rate in", values, "rate");
             return (Criteria) this;
         }
 
-        public Criteria andRateNotIn(List<BigDecimal> values) {
+        public Criteria andRateNotIn(List<Float> values) {
             addCriterion("rate not in", values, "rate");
             return (Criteria) this;
         }
 
-        public Criteria andRateBetween(BigDecimal value1, BigDecimal value2) {
+        public Criteria andRateBetween(Float value1, Float value2) {
             addCriterion("rate between", value1, value2, "rate");
             return (Criteria) this;
         }
 
-        public Criteria andRateNotBetween(BigDecimal value1, BigDecimal value2) {
+        public Criteria andRateNotBetween(Float value1, Float value2) {
             addCriterion("rate not between", value1, value2, "rate");
             return (Criteria) this;
         }
